requirements.txt
Jinja2==3.1.3
PyYAML==6.0.1
pytest==8.1.1
requests
feedparser
